+++ CHANGELOG.md
@@ -2,6 +2,7 @@ v.13.0-rc.1
 =============
 
 * Switched `kbox status` on `linux` to use PID file existence instead of `service` wrapper. [#1539](https://github.com/kalabox/kalabox/issues/1539)
+* Provided a check to ensure the `docker` daemon is up on Windows and OSX. [#1387](https://github.com/kalabox/kalabox/issues/1387)
 
 v.13.0-beta.4
 =============